requirements.txt
Django==1.10.3
django-filter==1.0.0
djangorestframework==3.5.3
Markdown==2.6.7
gunicorn==19.6.0
psycopg2==2.6.2
redis==2.10.5
pysha3==1.0.1; python_version < '3.6'